Refactor namespace to a single file
