+++ tests/core/namespace/test_pydantic_path.py
@@ -1,67 +1,64 @@
-from pydantic import ValidationError
 from pytest import raises
 
-from foundry.core.namespace.Namespace import Namespace
-from foundry.core.namespace.Path import Path
-from foundry.core.namespace.PydanticPath import PydanticPath
+from foundry.core.namespace import Namespace, Path
 
 
 def test_pydantic_path_name_empty():
-    assert Path() == PydanticPath(use_parent=False, path="").to_path()
+    assert Path() == Path.validate(use_parent=False, path="")
 
 
 def test_pydantic_path_short_simple():
-    assert Path.from_string("a") == PydanticPath(use_parent=False, path="a").to_path()
+    assert Path.from_string("a") == Path.validate(use_parent=False, path="a")
 
 
 def test_pydantic_path_simple():
-    assert Path.from_string("hello") == PydanticPath(use_parent=False, path="hello").to_path()
+    assert Path.from_string("hello") == Path.validate(use_parent=False, path="hello")
 
 
 def test_pydantic_path_simple_caps():
-    assert Path.from_string("HELLO") == PydanticPath(use_parent=False, path="HELLO").to_path()
+    assert Path.from_string("HELLO") == Path.validate(use_parent=False, path="HELLO")
 
 
 def test_pydantic_path_with_underscore_start():
-    assert Path.from_string("_hello") == PydanticPath(use_parent=False, path="_hello").to_path()
+    assert Path.from_string("_hello") == Path.validate(use_parent=False, path="_hello")
 
 
 def test_pydantic_path_with_dot():
-    assert Path.from_string("hello.world") == PydanticPath(use_parent=False, path="hello.world").to_path()
+    assert Path.from_string("hello.world") == Path.validate(use_parent=False, path="hello.world")
 
 
 def test_pydantic_path_with_incomplete_dot():
-    with raises(ValidationError):
-        PydanticPath(use_parent=False, path="hello.world.")
+    with raises(ValueError):
+        Path.validate(use_parent=False, path="hello.world.")
 
 
 def test_pydantic_path_with_invalid_string():
-    with raises(ValidationError):
-        PydanticPath(use_parent=False, path="hello@world")
+    with raises(ValueError):
+        Path.validate(use_parent=False, path="hello@world")
 
 
 def test_pydantic_path_with_invalid_middle_name():
-    with raises(ValidationError):
-        PydanticPath(use_parent=False, path="hello..world")
+    with raises(ValueError):
+        Path.validate(use_parent=False, path="hello..world")
 
 
 def test_pydantic_path_no_parent():
-    with raises(ValidationError):
-        PydanticPath(path="foo")
+    with raises(ValueError):
+        Path.validate(path="foo")
 
 
 def test_pydantic_path_namespace_validation_root():
     namespace = Namespace(children={"foo": Namespace()})
 
-    assert Path() == PydanticPath(parent=namespace, path="").to_path()
+    assert Path() == Path.validate(parent=namespace, path="")
 
 
 def test_pydantic_path_namespace_validation_simple():
     namespace = Namespace(children={"foo": Namespace()})
 
-    assert Path.from_string("foo") == PydanticPath(parent=namespace, path="foo").to_path()
+    assert Path.from_string("foo") == Path.validate(parent=namespace, path="foo")
 
 
 def test_pydantic_path_int():
-    with raises(ValidationError):
-        PydanticPath(use_parent=False, path=1)  # type: ignore
+    with raises(TypeError):
+        Path.validate(use_parent=False, path=1)  # type: ignore
